AcmeHrApp/backend/graph_ql/all_type_defs/get_paginated_colleagues.py
from graphene import relay, ObjectType, Boolean
from .login import ColleaguesConnection

class PaginatedColleaguesType(ObjectType):
    hasNextPage = Boolean()
    colleagues = relay.ConnectionField(ColleaguesConnection)
    