AcmeHrApp/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
whitenoise==6.0.0
django-environ==0.8.1
gunicorn==20.1.0
psycopg2==2.9.3
unittest-xml-reporting==3.0.2
coverage==6.3.2
graphene-django==3.0.0b7
django-cors-headers==3.11.0
